Make exposure time scale factor work.

+++ src/printjob.h
@@ -9,10 +9,10 @@ public:
     QString slicedSvgFileName;
     QString pngFilesPath;
     int layerCount;
-    int layerThickness;          // µm
-    double exposureTime;         // s
-    int exposureTimeScaleFactor; // for first two layers
-    int powerLevel;              // 0..255
+    int layerThickness;             // µm
+    double exposureTime;            // s
+    double exposureTimeScaleFactor; // for first two layers
+    int powerLevel;                 // 0..255
 
 };
